+++ jest.config.js
@@ -1,5 +1,4 @@
 module.exports = {
   preset: 'ts-jest',
-  testEnvironment: 'node',
-  coveragePathIgnorePatterns: ["<rootDir>/src/contracts"]
+  testEnvironment: 'node'
 }
